requirements.txt
annotated-types==0.7.0
anyio==4.11.0
beautifulsoup4==4.14.2
certifi==2025.10.5
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Markdown==3.9
ollama==0.6.0
pydantic==2.12.3
pydantic_core==2.41.4
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.1
sniffio==1.3.1
soupsieve==2.8
typing-inspection==0.4.2
typing_extensions==4.15.0
pyqt5
ollama
beautifulsoup4
markdown
pyserial
matplotlib
pyqtgraph
